Move GraphQL types purely for directive arguments into different directory
