+++ packages/graphql/src/graphql/directives/arguments/enums/CallbackOperation.ts
