+++ packages/graphql/src/graphql/directives/relationship.ts
@@ -18,8 +18,8 @@
  */
 
 import { DirectiveLocation, GraphQLDirective, GraphQLNonNull, GraphQLString } from "graphql";
-import { RelationshipDirectionEnum } from "../enums/RelationshipDirection";
-import { RelationshipQueryDirectionEnum } from "../enums/RelationshipQueryDirection";
+import { RelationshipDirectionEnum } from "./arguments/enums/RelationshipDirection";
+import { RelationshipQueryDirectionEnum } from "./arguments/enums/RelationshipQueryDirection";
 import { RelationshipQueryDirectionOption } from "../../constants";
 
 export const relationshipDirective = new GraphQLDirective({
